refactor: Move the classes in the BinaryInjection project into...
their proper namespace based on folders, instead of keeping everything in the CoreHook.BinaryInjection namespace
