+++ src/CoreHook.BinaryInjection/BinaryLoader/Memory/IMemoryManager.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Text;
 
-namespace CoreHook.BinaryInjection
+namespace CoreHook.BinaryInjection.BinaryLoader.Memory
 {
     public interface IMemoryManager : IDisposable
     {
