+++ src/CoreHook.ManagedHook/Remote/RemoteHooking.cs
@@ -6,6 +6,10 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization.Formatters.Binary;
 using CoreHook.BinaryInjection;
+using CoreHook.BinaryInjection.BinaryLoader;
+using CoreHook.BinaryInjection.BinaryLoader.Memory;
+using CoreHook.BinaryInjection.BinaryLoader.Windows;
+using CoreHook.BinaryInjection.Host;
 using CoreHook.CoreLoad;
 using CoreHook.IPC.Platform;
 using CoreHook.ManagedHook.ProcessUtils;
